feat: create middleware to set user auth status
This middleware sets the user auth status to the request context with
the key of type contextKey (check session.go to see the type definition
for it).

In other words, it checks the presence of the userID session key and if
the user with that id exists in the database.

If the key exists and there's a user with that id, then add a status of
true to the request context with the key contextKeyIsUserAuthenticated.